Enhance the alarm metrics by providing additional details
Opening this PR in order to add another collector, ``junos_alarm_set``
which provides further information for the alarms currently set on the
device: type, class and description. Using this information, one is able
to configure alerts on specific chassis or system alarms, as the number
of yellow / red alarms is only useful in dashboards.