Collect metrics on the IS-IS adjacency state
The current IS-IS metrics only provide the total counters, i.e., number
of adjacencies and how many are up. This is great for a dashboard-like
approach, but insufficient when you need to look into a more granular
per-adjacency, for example for alerting.
Adding a new ``junos_isis_adjacency_state`` metrics family that provide
the adjacency state.